print top 30

+++ run.py
@@ -152,35 +152,41 @@ with tf.Session() as sess:
 
 	elif FLAGS.mode == 'sens_anlys':
 		deco_print('Executing Sensitivity Analysis Mode\n')
-		count = np.zeros(shape=(5,), dtype=int)
-		gradients = np.zeros(shape=(5, model._config.num_category, model._config.feature_dim), dtype=float)
-		epoch_start = time.time()
-		cur_epoch_step = 0
-		sample_step = 0
-		for _, (x, y, info, x_cur) in enumerate(dl.iterate_one_epoch(model._config.batch_size, output_current_status=True)):
-			if sample_step != FLAGS.sample_size:
-				count += np.sum(x_cur, axis=0)
-				feed_dict = {model._x_placeholder:x, model._y_placeholder:y}
-				gradients_i, = sess.run(fetches=[model._x_gradients], feed_dict=feed_dict)
-				for v in range(model._config.num_category):
-					gradients_i_v = gradients_i[v]
-					gradients[:,v,:] += x_cur.T.dot(np.absolute(gradients_i_v))
-				sample_step += 1
-			if info['epoch_step'] != cur_epoch_step:
-				epoch_last = time.time() - epoch_start
-				time_est = epoch_last / (info['idx_file'] + 1) * info['num_file']
-				deco_print('Elapse / Estimate: %.2fs / %.2fs     ' %(epoch_last, time_est), end='\r')
-				cur_epoch_step = info['epoch_step']
-				sample_step = 0
-		gradients /= count[:, np.newaxis, np.newaxis]
-		deco_print('Saving Output in %s' %os.path.join(FLAGS.logdir, 'ave_absolute_gradient.npy'))
-		np.save(os.path.join(FLAGS.logdir, 'ave_absolute_gradient.npy'), gradients)
+
+		if not os.path.exists(os.path.join(FLAGS.logdir, 'ave_absolute_gradient.npy')):
+			count = np.zeros(shape=(5,), dtype=int)
+			gradients = np.zeros(shape=(5, model._config.num_category, model._config.feature_dim), dtype=float)
+			epoch_start = time.time()
+			cur_epoch_step = 0
+			sample_step = 0
+			for _, (x, y, info, x_cur) in enumerate(dl.iterate_one_epoch(model._config.batch_size, output_current_status=True)):
+				if sample_step != FLAGS.sample_size:
+					count += np.sum(x_cur, axis=0)
+					feed_dict = {model._x_placeholder:x, model._y_placeholder:y}
+					gradients_i, = sess.run(fetches=[model._x_gradients], feed_dict=feed_dict)
+					for v in range(model._config.num_category):
+						gradients_i_v = gradients_i[v]
+						gradients[:,v,:] += x_cur.T.dot(np.absolute(gradients_i_v))
+					sample_step += 1
+				if info['epoch_step'] != cur_epoch_step:
+					epoch_last = time.time() - epoch_start
+					time_est = epoch_last / (info['idx_file'] + 1) * info['num_file']
+					deco_print('Elapse / Estimate: %.2fs / %.2fs     ' %(epoch_last, time_est), end='\r')
+					cur_epoch_step = info['epoch_step']
+					sample_step = 0
+			gradients /= count[:, np.newaxis, np.newaxis]
+			deco_print('Saving Output in %s' %os.path.join(FLAGS.logdir, 'ave_absolute_gradient.npy'))
+			np.save(os.path.join(FLAGS.logdir, 'ave_absolute_gradient.npy'), gradients)
+
+		deco_print('Top 30:')
+		top_covariate = feature_ranking(FLAGS.logdir, dl._idx2covariate)
+		for item in top_covariate:
+			print(item)
 		deco_print('Sensitivity Analysis Finished')
 
 	elif FLAGS.mode == 'sens_anlys_pair':
 		deco_print('Executing Sensitivity Analysis (Pairs) Mode')
 		deco_print('Consider Real-Valued Pairs Only\n')
-		count = np.zeros(shape=(5,), dtype=int)
 
 		### real-valued pairs
 		num_feature_pairs = dl._covariate_count_float * (dl._covariate_count_float - 1) // 2
@@ -189,6 +195,7 @@ with tf.Session() as sess:
 		###
 
 		if not os.path.exists(os.path.join(FLAGS.logdir, 'ave_absolute_gradient_2.npy')):
+			count = np.zeros(shape=(5,), dtype=int)
 			gradients = np.zeros(shape=(5, model._config.num_category, num_feature_pairs))
 			epoch_start = time.time()
 			cur_epoch_step = 0
